+++ mitra_activitypub/src/identity.rs
@@ -50,7 +50,8 @@ pub fn create_identity_claim_fep_c390(
     let message = match proof_type {
         IdentityProofType::LegacyEip191IdentityProof
             | IdentityProofType::LegacyMinisignIdentityProof
-            => unimplemented!("expected FEP-c390 compatible proof type"),
+            | IdentityProofType::FepC390LegacyJcsEddsaProof
+            => unimplemented!("expected supported proof type"),
         IdentityProofType::FepC390JcsBlake2Ed25519Proof => {
             subject.as_did_key().expect("did:key should be used");
             canonicalize_object(&claim)?
@@ -58,16 +59,6 @@ pub fn create_identity_claim_fep_c390(
         IdentityProofType::FepC390JcsEip191Proof => {
             subject.as_did_pkh().expect("did:pkh should be used");
             canonicalize_object(&claim)?
-        },
-        IdentityProofType::FepC390LegacyJcsEddsaProof => {
-            subject.as_did_key().expect("did:key should be used");
-            #[allow(deprecated)]
-            let proof_config = IntegrityProofConfig::jcs_eddsa_legacy(
-                &subject.to_string(),
-                proof_created_at,
-            );
-            let hash_data = prepare_jcs_sha256_data(&claim, &proof_config)?;
-            hex::encode(hash_data)
         },
         IdentityProofType::FepC390EddsaJcsProof => {
             subject.as_did_key().expect("did:key should be used");
@@ -102,7 +93,8 @@ pub fn create_identity_proof_fep_c390(
     let integrity_proof = match proof_type {
         IdentityProofType::LegacyEip191IdentityProof
             | IdentityProofType::LegacyMinisignIdentityProof
-            => unimplemented!("expected FEP-c390 compatible proof type"),
+            | IdentityProofType::FepC390LegacyJcsEddsaProof
+            => unimplemented!("expected supported proof type"),
         IdentityProofType::FepC390JcsBlake2Ed25519Proof => {
             let did_key = subject.as_did_key()
                 .expect("did:key should be used");
@@ -112,16 +104,6 @@ pub fn create_identity_proof_fep_c390(
             let did_pkh = subject.as_did_pkh()
                 .expect("did:pkh should be used");
             IntegrityProof::jcs_eip191(did_pkh, signature_bin)
-        },
-        IdentityProofType::FepC390LegacyJcsEddsaProof => {
-            let did_key = subject.as_did_key()
-                .expect("did:key should be used");
-            #[allow(deprecated)]
-            let proof_config = IntegrityProofConfig::jcs_eddsa_legacy(
-                &did_key.to_string(),
-                proof_created_at,
-            );
-            IntegrityProof::new(proof_config, signature_bin)
         },
         IdentityProofType::FepC390EddsaJcsProof => {
             let did_key = subject.as_did_key()
@@ -166,7 +148,6 @@ mod tests {
             ecdsa_public_key_to_address_hex,
         },
         json_signatures::{
-            create::sign_object_eddsa,
             proofs::ProofType,
             verify::{
                 get_json_signature,
@@ -232,64 +213,6 @@ mod tests {
         let result = verify_eip191_json_signature(
             &did_pkh,
             &signature_data.object,
-            &signature_data.signature,
-        );
-        assert_eq!(result.is_ok(), true);
-    }
-
-    #[test]
-    #[allow(deprecated)]
-    fn test_create_and_verify_identity_proof_eddsa_legacy() {
-        // jcs-eddsa-2022 (minisign-unhashed); no context injection
-        let did_str = "did:key:z6MkrJVnaZkeFzdQyMZu1cgjg7k1pZZ6pvBQ7XJPt4swbTQ2";
-        let did = did_str.parse().unwrap();
-        let secret_key_multibase = "z3u2en7t5LR2WtQH5PfFqMqwVHBeXouLzo6haApm8XHqvjxq";
-        let secret_key = ed25519_secret_key_from_multikey(secret_key_multibase).unwrap();
-        let actor_id = "https://server.example/users/alice";
-        let created_at = DateTime::parse_from_rfc3339("2023-02-24T23:36:38Z")
-            .unwrap().with_timezone(&Utc);
-        let (claim, _) = create_identity_claim_fep_c390(
-            actor_id,
-            &did,
-            &IdentityProofType::FepC390LegacyJcsEddsaProof,
-            created_at,
-        ).unwrap();
-        let claim_value = serde_json::to_value(claim).unwrap();
-        let identity_proof = sign_object_eddsa(
-            &secret_key,
-            &did.to_string(),
-            &claim_value,
-            Some(created_at),
-            true, // legacy cryptosuite
-            false, // no proof context
-            false, // context injection not required
-        ).unwrap();
-        let expected_result = json!({
-            "type": "VerifiableIdentityStatement",
-            "subject": "did:key:z6MkrJVnaZkeFzdQyMZu1cgjg7k1pZZ6pvBQ7XJPt4swbTQ2",
-            "alsoKnownAs": "https://server.example/users/alice",
-            "proof": {
-                "type": "DataIntegrityProof",
-                "cryptosuite": "jcs-eddsa-2022",
-                "verificationMethod": "did:key:z6MkrJVnaZkeFzdQyMZu1cgjg7k1pZZ6pvBQ7XJPt4swbTQ2",
-                "proofPurpose": "assertionMethod",
-                "proofValue": "zYqr4eFzrnUWiBDaa7SmBhfaSBiv6BFRsDRGkmaCJpXArPBspFWNM6NXu77R7JakdzbUdjZihBa28LuWscZxSfRk",
-                "created": "2023-02-24T23:36:38Z",
-            },
-        });
-        assert_eq!(identity_proof, expected_result);
-
-        let signature_data = get_json_signature(&identity_proof).unwrap();
-        assert_eq!(
-            signature_data.proof_type,
-            ProofType::JcsEddsaSignature,
-        );
-        let public_key = did.as_did_key().unwrap()
-            .try_ed25519_key().unwrap();
-        let result = verify_eddsa_json_signature(
-            &public_key,
-            &signature_data.object,
-            &signature_data.proof_config,
             &signature_data.signature,
         );
         assert_eq!(result.is_ok(), true);